backend/search/requirements.txt
# Core Dependencies
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.4.2
python-dotenv==1.0.0

# Database
motor==3.3.1
pymongo==4.6.0

# Search & ML
faiss-cpu==1.7.4
numpy==1.24.3
scikit-learn==1.3.2
torch==2.1.0
transformers==4.34.0

# Caching & Performance
redis==5.0.1
aioredis==2.0.1

# Monitoring & Logging
elasticsearch==8.10.0
prometheus-client==0.17.1
opentelemetry-api==1.20.0
opentelemetry-sdk==1.20.0
structlog==23.2.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
httpx==0.25.0

# Development
black==23.10.0
isort==5.12.0
flake8==6.1.0
mypy==1.6.1 